New tests and major fixes
